+++ src/arcemu-world/Guild.cpp
@@ -557,7 +557,7 @@ bool Guild::LoadFromDB(Field * f)
 		{
 			GuildLogEvent * li = new GuildLogEvent;
 			li->iLogId = result->Fetch()[0].GetUInt32();
-			li->iEvent = result->Fetch()[3].GetUInt32();
+			li->iEvent = static_cast<uint8>( result->Fetch()[3].GetUInt32() );
 			li->iTimeStamp = result->Fetch()[2].GetUInt32();
 			li->iEventData[0] = result->Fetch()[4].GetUInt32();
 			li->iEventData[1] = result->Fetch()[5].GetUInt32();
